NotesForAll/requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-phonenumber-field==4.0.0
django-twilio==0.13.0.post1
idna==2.10
phonenumbers==8.12.6
pkg-resources==0.0.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.44.0
urllib3==1.25.9
